move volume view to content app
