+++ ohmg/content/views.py
@@ -1,6 +1,9 @@
+import json
 import logging
+from datetime import datetime
 
 from django.conf import settings
+from django.http import JsonResponse
 from django.shortcuts import render, get_object_or_404
 from django.views import View
 from django.middleware import csrf
@@ -11,12 +14,77 @@ from ohmg.georeference.models import (
     Document,
     ItemBase,
 )
-
-from ohmg.loc_insurancemaps.models import find_volume
+from ohmg.loc_insurancemaps.models import Volume, find_volume
+from ohmg.loc_insurancemaps.tasks import load_docs_as_task
+from ohmg.frontend.context_processors import user_info_from_request
 
 logger = logging.getLogger(__name__)
 
-class ResourceView(View):
+
+class ItemView(View):
+
+    def get(self, request, volumeid):
+
+        volume = get_object_or_404(Volume, pk=volumeid)
+        volume_json = volume.serialize(include_session_info=True)
+
+        context_dict = {
+            "svelte_params": {
+                "TITILER_HOST": settings.TITILER_HOST,
+                "VOLUME": volume_json,
+                "CSRFTOKEN": csrf.get_token(request),
+                "USER": user_info_from_request(request),
+                "MAPBOX_API_KEY": settings.MAPBOX_API_TOKEN,
+            }
+        }
+        return render(
+            request,
+            "content/item.html",
+            context=context_dict
+        )
+
+    def post(self, request, volumeid):
+
+        body = json.loads(request.body)
+        operation = body.get("operation", None)
+
+        if operation == "initialize":
+            volume = Volume.objects.get(pk=volumeid)
+            if volume.loaded_by is None:
+                volume.loaded_by = request.user
+                volume.load_date = datetime.now()
+                volume.save(update_fields=["loaded_by", "load_date"])
+            load_docs_as_task.delay(volumeid)
+            volume_json = volume.serialize(include_session_info=True)
+            volume_json["status"] = "initializing..."
+
+            return JsonResponse(volume_json)
+
+        elif operation == "set-index-layers":
+
+            volume = Volume.objects.get(pk=volumeid)
+
+            lcat_lookup = body.get("layerCategoryLookup", {})
+
+            for cat in volume.sorted_layers:
+                volume.sorted_layers[cat] = [k for k, v in lcat_lookup.items() if v == cat]
+
+            volume.save(update_fields=["sorted_layers"])
+            volume_json = volume.serialize(include_session_info=True)
+            return JsonResponse(volume_json)
+
+        elif operation == "refresh":
+            volume = Volume.objects.get(pk=volumeid)
+            volume_json = volume.serialize(include_session_info=True)
+            return JsonResponse(volume_json)
+
+        elif operation == "refresh-lookups":
+            volume = Volume.objects.get(pk=volumeid)
+            volume.refresh_lookups()
+            volume_json = volume.serialize(include_session_info=True)
+            return JsonResponse(volume_json)
+
+class VirtualResourceView(View):
 
     def get(self, request, pk):
 
@@ -28,7 +96,6 @@ class ResourceView(View):
 
         split_summary = resource.get_split_summary()
         georeference_summary = resource.get_georeference_summary()
-        sessions_json = resource.get_sessions(serialize=True)
         resource_json = resource.serialize()
 
         volume = find_volume(resource)
@@ -45,11 +112,9 @@ class ResourceView(View):
                     'RESOURCE': resource_json,
                     'VOLUME': volume_json,
                     'CSRFTOKEN': csrf.get_token(request),
-                    'USER_AUTHENTICATED': request.user.is_authenticated,
-                    'USER_STAFF': request.user.is_staff,
+                    "USER": user_info_from_request(request),
                     "SPLIT_SUMMARY": split_summary,
                     "GEOREFERENCE_SUMMARY": georeference_summary,
-                    "SESSION_HISTORY": sessions_json,
                     "MAPBOX_API_KEY": settings.MAPBOX_API_TOKEN,
                     "OHMG_API_KEY": settings.OHMG_API_KEY,
                     "SESSION_API_URL": reverse("api-beta:session_list"),
